(WIP) today's work rewriting template uri handling methods
